fix Core Data

+++ GuessTheTuneGame/GuessTheTuneGame/ViewControllers/FavouritesViewController.swift
@@ -10,24 +10,20 @@ import CoreData
 
 class FavouritesViewController: UIViewController, UITableViewDelegate, UITableViewDataSource
 {
-    let tableView: UITableView =
-        {
-            let table = UITableView()
-            table.register(UITableViewCell.self,forCellReuseIdentifier: "cell")
-            
-            return table
-        }()
+    private var tableView: UITableView?
+    private var favTracks: [TrackItem]?
+    var managedObjectContext: NSManagedObjectContext?
     
-   private var favTracks = [TrackItem]()
-    
-    override func loadView()
+    override func viewDidLoad()
     {
-        super.loadView()
-        self.view.tintColor = CustomColor.customGreen
-        self.title = "Favourite Tracks"
+        super.viewDidLoad()
+        self.tableView = UITableView(frame: CGRect(x: 0, y: 0, width: view.frame.width, height: view.frame.height), style: .plain)
+        view.tintColor = CustomColor.customGreen
+        title = "Favourite Tracks"
         
         getAllItems()
         configureTableView()
+        print("favTracks size: ", favTracks?.count)
     }
     
     init()
@@ -38,6 +34,19 @@ class FavouritesViewController: UIViewController, UITableViewDelegate, UITableVi
     required init?(coder: NSCoder)
     {
         fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func configureTableView()
+    {
+        print("configureTableView()")
+        guard let tableView = tableView else {return}
+
+        tableView.register(UITableViewCell.self,forCellReuseIdentifier: "cell")
+        tableView.delegate = self
+        tableView.dataSource = self
+        tableView.frame = view.bounds
+        tableView.backgroundColor = CustomColor.offWhite
+        view.addSubview(tableView)
     }
     
     //MARK: TableView Methods
@@ -51,32 +60,34 @@ class FavouritesViewController: UIViewController, UITableViewDelegate, UITableVi
     {
         if (editingStyle == .delete)
         {
-            CoreData_.deleteItem(track: favTracks[indexPath.row])
+            if let favTracks = favTracks
+            {
+                do
+                {
+                    try CoreData_.deleteItem(track: favTracks[indexPath.row])
+                }
+                catch
+                {
+                    print("tableView edit error")
+                }
+            }
             getAllItems()
             tableView.reloadData()
         }
     }
     
-    func configureTableView()
-    {
-        tableView.delegate = self
-        tableView.dataSource = self
-        tableView.frame = view.bounds
-        tableView.backgroundColor = CustomColor.offWhite
-        view.addSubview(tableView)
-             
-    }
-    
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int
     {
+        guard let favTracks = favTracks else {return 0}
         return favTracks.count
-       
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell
     {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
         
+        guard let favTracks = favTracks else {return UITableViewCell()}
+
         let tracks = favTracks[indexPath.row]
         cell.textLabel?.text = tracks.name
         cell.textLabel?.font = UIFont(name:"HelveticaNeue-Bold", size:18)
@@ -92,22 +103,20 @@ class FavouritesViewController: UIViewController, UITableViewDelegate, UITableVi
     
     //MARK: Get favourite tracks
     
-    func getAllItems() -> Void
+    private func getAllItems()
     {
         do
         {
-            favTracks = try context.fetch(TrackItem.fetchRequest())
+            self.favTracks = try managedObjectContext?.fetch(TrackItem.fetchRequest())
+            guard let tableView = tableView else {return}
+
             DispatchQueue.main.async{
-                self.tableView.reloadData()
+                tableView.reloadData()
             }
         }
         catch
         {
-            //throw exception
             print("getAllItems: Error")
         }
-        
     }
-    
-
 }
